\name{add.mutations}
\Rdversion{0.1}
\alias{add.mutations}
\title{Add Mutations on a Coalescent Tree}
\description{
Add mutations on a coalescent tree.
}
\usage{
  add.mutations(tree,mu,graphics = TRUE, labels = TRUE)
}
\arguments{
  \item{tree}{an object generated by the \code{\link{simulate.tree}} command}
  \item{mu}{mutation rate}
  \item{graphics}{a logical variable, which is TRUE if the user wants to plot the history, and FALSE otherwise}
  \item{labels}{a logical variable, which is TRUE if the user wants labels to be plotted on the tree, and FALSE otherwise}
}
\details{
Adds mutation on a coalescent tree generated by the \code{\link{simulate.tree}} command, using the mutation rate \sQuote{mu}. The mutation model is an Infinite Allele Model. If \sQuote{graphics = TRUE}, then a coalescent tree is drawn, with mutations sketched on the tree. Allelic states of the leafs are represented with different colors. If \sQuote{labels = TRUE}, then the labels are drawn on the graph.
}
\value{
an object of class \sQuote{history} which is a list with components:
\item{state}{the allelic state of each leaf}
\item{node}{the list of nodes (internal and leafs) in the tree}
\item{ancestor}{the ancestor of each node}
\item{label}{the label of each leaf}
\item{branch.length}{the length of the branch above each node}
}
\examples{
## This is to simulate a coalescent tree using the Hudson's algorithm, 
## with 20 sampled lineages. In this example, the current population size
## is 100. It was 1000 before a bottleneck that occurred 50 generations ago

tree <- simulate.tree(method = "hudson",sample = 20,current = 100,ancestral = 1000,time = 50)

## This is to generate a history, with a graphical outputs with labels, using a mutation rate of 0.001

history <- add.mutations(tree,mu = 0.001,graphics = TRUE, labels = TRUE)
}
