\name{coalesceR}
\Rdversion{0.2}
\alias{coalesceR}
\title{A Pedagogical Tool to Teach the Basics of Coalescent Theory}
\description{Coalescent theory provides a powerful modeling tool for population genetics.
It is based on a genealogical interpretation of the history of gene lineages in populations.
Although classical population genetics theory considers the evolution of gene lineages forward in time,
coalescent theory considers the genealogical history of a sample of genes, looking backward in time.
In the absence of selection (i.e., if all genetic variants are selectively neutral),
then the mutation process can be decoupled from the genealogical process. Therefore,
selectively neutral mutations can then be superimposed on the genealogical history.
This package is intended to  provide a pedagogical tool to teach the basics of coalescent theory.
It enables to simulate gene genealogies in an  isolated population of constant effective size, or in a population that changed in size some generations in the past.
Gene genealogies can be simulated using a continuous time approximation, or a generation-by-generation algorithm.Graphical outputs show the consequence of population size changes on the shape of genealogies.
Mutations can then be added on the gene genealogies, and summary statistics (e.g., heterozygosity, number of alleles) can be computed from gene histories.
}
\author{Renaud Vitalis \email{vitalis@supagro.inra.fr}}
