\name{draw.tree}
\Rdversion{0.2}
\alias{draw.tree}
\title{Draw a Coalescent Tree}
\description{
		Draw a coalescent tree.
}
\usage{
	draw.tree(tree,labels = TRUE)
}
\arguments{
	\item{tree}{a tree object generated by the \code{\link{sim.tree}} command}
	\item{labels}{a logical variable, which is TRUE if the user wants labels to be plotted on the tree, and FALSE otherwise}
}
\details{
	Once the \code{\link{sim.tree}} command line has been executed, \code{\link{draw.tree}} can be used to draw the corresponding coalescent tree
}
\value{
	Draws a coalescent tree generated by the \code{\link{sim.tree}} command. If \sQuote{labels = TRUE}, then the labels are drawn on the graph.
}
\examples{
## This is to simulate a coalescent tree using the Hudson's algorithm, with 20 sampled lineages.
## In this example, the current population is 100. It was 1000 before a bottleneck that occurred 50 generations ago

tree <- sim.tree(method = "hudson",sample = 20,current = 100,ancestral = 1000,time = 50)

## This plots (on the screen) the coalescent tree, with labels

draw.tree(tree,labels = TRUE)
}

