\name{heterozygosity}
\Rdversion{0.2}
\alias{heterozygosity}
\title{Compute the Heterozygosity of a Sample of Genes}
\description{
Compute Nei's unbiased heterozygosity of a sample of genes.
}
\usage{
  heterozygosity(tree.with.mutations)
}
\arguments{
  \item{tree.with.mutations}{an object generated by the \code{\link{add.mutations}} command}
}
\details{
Once the \code{\link{sim.tree}} and \code{\link{add.mutations}} command lines have been executed, \code{\link{heterozygosity}} can be used to compute Nei's heterozygosity of the sample}
\value{
	Nei's unbiased heterozygosity of the sample contained in the object \sQuote{history}
}
\examples{
## This is to simulate a coalescent tree using the Hudson's algorithm, 
## with 20 sampled lineages. In this example, the current population size
## is 100. It was 1000 before a bottleneck that occurred 50 generations ago

tree <- sim.tree(method = "hudson",sample = 20,current = 100,ancestral = 1000,time = 50)

## This is to generate a history, with a graphical outputs with labels, using a mutation rate of 0.001

history <- add.mutations(tree,mu = 0.001,graphics = TRUE, labels = TRUE)

## This is to compute Nei's unbiased heterozygosity for that sample

heterozygosity(history)
}
