\name{plot.tree}
\Rdversion{0.1}
\alias{plot.tree}
\title{Plot a Coalescent Tree}
\description{
Plot a coalescent tree.
}
\usage{
  plot.tree(tree,labels = TRUE)
}
\arguments{
  \item{tree}{a tree object generated by the \code{\link{simulate.tree}} command}
  \item{labels}{a logical variable, which is TRUE if the user wants labels to be plotted on the tree, and FALSE otherwise}
}
\details{
}
\value{
Plots a coalescent tree generated by the \code{\link{simulate.tree}} command. If \sQuote{labels = TRUE}, then the labels are drawn on the graph.
}
\examples{
## This is to simulate a coalescent tree using the Hudson's algorithm, with 20 sampled lineages.
## In this example, the current population is 100. It was 1000 before a bottleneck that occurred 50 generations ago

tree <- simulate.tree(method = "hudson",sample = 20,current = 100,ancestral = 1000,time = 50)

## This plots (on the screen) the coalescent tree, with labels

plot.tree(tree,labels = TRUE)
}

