\name{TMRCA}
\Rdversion{0.1}
\alias{TMRCA}
\title{Compute the TMRCA of a Coalescent Tree}
\description{
Compute the Time to the Most Recent Common Ancestor (TMRCA) of a coalescent tree.
}
\usage{
  TMRCA(tree)
}
\arguments{
  \item{tree}{an object generated by the \code{\link{simulate.tree}} command}
}
\details{
Once the \code{\link{simulate.tree}} command line has been executed, \code{\link{TMRCA}} can be used to compute the TMRCA of the tree. Note that the \code{\link{TMRCA}} command may also be used on an object of class \sQuote{history}
}
\value{
	TMRCA of the coalescent tree contained in the object \sQuote{tree}
}
\examples{
## This is to simulate a coalescent tree using the Hudson's algorithm, 
## with 20 sampled lineages. In this example, the current population size
## is 100. It was 1000 before a bottleneck that occurred 50 generations ago

tree <- simulate.tree(method = "hudson",sample = 20,current = 100,ancestral = 1000,time = 50)

## This is to compute the TMRCA of that tree
TMRCA(tree)
}
